requirements.txt
## -----------------------------------
#  python pip requirements
## -----------------------------------


##
#-e .[docs,test]


# Toolsets
pycodestyle
invoke

# Base
dxlclient >= 3.1.0.586
dxltieclient >= 0.1.0

named_constants==1.0
PyYAML
argparse
pathtools
urllib3
requests==2.7.0
watchdog==0.8.3
